added new hood useOffersRequests
